college/multiagentapp/ml_workflow/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0  # For Excel file support
streamlit>=1.22.0
plotly>=5.13.0
groq==0.4.0  # For Groq API
python-dotenv==1.0.0  # For environment variables
requests>=2.31.0  # For API calls 